requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
bleach==4.1.0
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-admin-inline-paginator==0.2
django-multiselectfield==0.1.12
django-qr-code==2.2.0
django-resized==0.3.11
django-summernote==0.8.11.6
django-weasyprint==1.1.0.post1
idna==3.2
marshmallow==3.12.1
mysqlclient==2.0.3
Pillow==8.2.0
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.5
